update prints and run black

+++ scripts/analysis/delete_old_branches.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-"""
+"""Prunes stale branches
 This script will look at all branches for a project and if the mode is delete, remove all branches that haven't been updated in 6 months.
 Expected environment variables:
     - GITLAB_URL: GitLab HTTPS instance URL
@@ -16,17 +16,31 @@ from dateutil.relativedelta import relativedelta
 import gitlab
 
 
-def evaluate_branches(branches, mode, diff_age) -> None:
+def get_open_merge_requests(project) -> tuple:
+    """
+    Returns a tuple of the source branch names of open merge requests within a project
+    """
+    open_merge_requests = project.mergerequests.list(all=True, state="opened")
+    open_mr_branches = [x.source_branch for x in open_merge_requests]
+    return tuple(open_mr_branches)
+
+
+def evaluate_branches(
+    branches,
+    open_mr_branches,
+    mode,
+    diff_age,
+) -> None:
     """
     Loop through GL project's list of branches to delete any that haven't been updated in the provided amount of time
-    Defaults to six months
+    Defaults to six (months)
     """
     diff_age_timestamp = (
         datetime.now() - relativedelta(months=int(diff_age))
     ).timestamp()
     stale_branch_count = 0
     branch_total = len(branches)
-    print(f"Total project branches: {branch_total}")
+    print(f"Searching for branches older than {diff_age} month(s)")
     for branch in branches:
         branch_name = branch.name
         committed_date = branch.commit["committed_date"]
@@ -36,31 +50,40 @@ def evaluate_branches(branches, mode, diff_age) -> None:
         last_updated = datetime.strftime(
             datetime.utcfromtimestamp(branch_update_age), "%Y-%m-%dT%H:%M:%SZ"
         )
-        print(f"Searching for branches older than {diff_age} month(s)")
-        # if the last update date is more than the defined amount of time, and not in tuple of preserved branches, delete the branch.
-        if diff_age_timestamp > branch_update_age and branch_name not in ("master"):
+        # if
+        #   the last update date is older (less than) than the defined amount of time
+        #   not in tuple of preserved branches
+        #   not in the tuple of open MR source branch names
+        # then delete the branch.
+        if (
+            diff_age_timestamp > branch_update_age
+            and branch_name not in ("master",)
+            and branch_name not in open_mr_branches
+        ):
             print(
                 f"""Stale branch:\t{branch_name}\n"""
-                f"""Last updated:\t{last_updated}\n"""
+                f"""Last updated:\t{last_updated}"""
             )
             stale_branch_count += 1
             if mode == "delete":
                 print("Deleting branch")
                 branch.delete()
+            print("")
         else:
             print(
                 f"""Retaining branch: {branch_name}\n"""
                 f"""Last updated:\t{last_updated}\n"""
             )
     print(
-        f"""Total stale branches:\t\t{stale_branch_count}\n"""
-        f"""Total retained branches:\t{branch_total - stale_branch_count}"""
+        f"""Total project branches:\t\t{branch_total}\n"""
+        f"""Total stale branches:\t\t{stale_branch_count}"""
     )
+    if mode == "delete":
+        print(f"Total retained branches:\t{branch_total - stale_branch_count}")
 
 
 def main() -> None:
-    """
-    Main method
+    """Main method
     use gitlab-python to retrieve branches and pass to delete function
     """
     url = os.environ["GITLAB_URL"]
@@ -73,9 +96,16 @@ def main() -> None:
 
     project = gl_instance.projects.get(project_id)
     branches = project.branches.list(all=True)
+    open_mr_branches = get_open_merge_requests(project)
 
     mode = os.environ.get("MODE", "dry_run")
-    evaluate_branches(branches, mode, diff_age)
+    print(f"Mode selected: {mode}")
+    evaluate_branches(
+        branches,
+        open_mr_branches,
+        mode,
+        diff_age,
+    )
 
 
 if __name__ == "__main__":
